+++ react/lib/JBMobileInput.tsx
@@ -15,19 +15,19 @@ export const JBMobileInput = forwardRef((props:Props, ref) => {
     () => element.current ?? undefined,
     [element],
   );
-  const {disabled,required,validationList,value,onBeforeinput,onBlur,onChange,onEnter,onFocus,onInput,onKeydown,onKeyup,...otherProps} = props;
-  useJBInputAttribute<JBMobileInputWebComponent>(element,{disabled,required,validationList,value,...otherProps});
+  const {disabled,initialValue,required,validationList,value,onBeforeinput,onBlur,onChange,onEnter,onFocus,onInput,onKeydown,onKeyup,...otherProps} = props;
+  useJBInputAttribute<JBMobileInputWebComponent>(element,{disabled,required,validationList,...otherProps});
   useJBInputEvents<JBMobileInputWebComponent>(element,{onBeforeinput,onBlur,onChange,onEnter,onFocus,onInput,onKeydown,onKeyup,...otherProps});
 
   return(
-    <jb-mobile-input ref={element} {...otherProps}>
+    <jb-mobile-input ref={element} value={value?.toString() ?? ""} initialValue={initialValue?.toString() ?? ""} {...otherProps}>
       {props.children}
     </jb-mobile-input>
   );
 });
 
 export type Props = BaseProps<JBMobileInputWebComponent> & {
-  // add special props here
+  initialValue?: string | number | null,
 }
 
 JBMobileInput.displayName = "JBMobileInput";