+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Added the React `initialValue` prop and forwarded `value` and `initialValue` directly as React 19 custom-element properties.
+
 ### Fixed
 
 - Added the missing `jb-input/react` build external and UMD global mapping.